requirements.txt
pydantic
pandas
fastapi
upstash_redis
psycopg2-binary
python-dotenv
requests
mysql-connector-python
uvicorn
